tui/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "localcode-tui"
version = "0.1.0"
description = "Textual TUI for LocalCode — vibe coding with local models"
requires-python = ">=3.10"
dependencies = [
    "textual>=1.0.0,<2",
    "websockets>=12.0,<14",
    "pyyaml>=6.0,<7",
    "rich>=13.0,<14",
    "httpx>=0.27,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "textual-dev>=1.0",
]

[project.scripts]
localcode-tui = "localcode_tui.app:main"
